Added code for calculating averag rating and propTypes in FeedbackStats
To calculate average, we have used the reduce function of array. It has two argumen one is the function and the other is the inital value of accumulator. The function has two args which are accumulator and current item. We used toFixed(1) for decimal precision and .replace to remove the 0 after decimal