Added router to the App
We use router to route the request as per url to the component. There are 2 types of it: BrowserRoute and HashRoute. In this case we have used BrowserRoute (and hence we don't need to use # in url). The Route should be wrapped within <Routes> and we mention the path as well as the element in the Route. The entire App JSX needs to be warpped inside <BroswerRouter> (or its alias name as we have done <Router>). We used exact in first case so that when /about is used the / component don't load.