Added Link of reacter-router-dom
react-router-dom library provides us with the Link Component which we can use in place of <a> tag. The sample syntax is <Link to="/path"> Text <Link/>. Also notice one thing that in HTML inspector we are seeing only those elements which are attached to that route and not the others.