Navigate component in react-router-dom
We can redirect the request coming to a route to another route based on condition by returing the Navigate component. We need to specify which component we need to Navigate to. In this case if the status is 404, the request coming to /post will be navigated/redirected to /notfound. In react-router-dom 5 it used to "Redirect" component.