useEffect() in react
We can useEffect() to introduce the side effect to the Component which will come into picture whenever there is change in a variable/object. If we want it to run on the component loading then we can pass the 2nd argument as empty array i.e. []. In this we are depending on editedFeedback to introduce the changes to the FeedbackForm component as well as RatingSelector component. Syntax of useEffect : useEffect(callback function,[objects/variables to be observed for change]);
We can use this to make http calls whenever a component loads.